#79 refactor(invitation) : 서비스 코드 리팩토링
