+++ src/main/java/com/coderder/colorMeeting/controller/InvitationController.java
@@ -2,7 +2,7 @@ package com.coderder.colorMeeting.controller;
 
 import com.coderder.colorMeeting.config.auth.PrincipalDetails;
 import com.coderder.colorMeeting.dto.request.TeamMemberRequestDto;
-import com.coderder.colorMeeting.service.InvitationService;
+import com.coderder.colorMeeting.service.InvitationServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/invite")
 public class InvitationController {
 
-    private final InvitationService invitationService;
+    private final InvitationServiceImpl invitationService;
 
     @RequestMapping(method = RequestMethod.GET)
     public ResponseEntity<?> showAllInvitations(@AuthenticationPrincipal PrincipalDetails userDetails) {
